+++ Screens/HomeScreen.tsx
@@ -9,30 +9,11 @@ import TabContent from "../src/Components/TabContent/TabContent";
 import OfferSwiper from "../src/Components/OfferGallery/OfferGallery";
 import slides from "../Data/Slides";
 import OfferSlides from "../Data/OfferSlides";
+import handleCardPress from "../logic/handleCardPress";
 
 const HomeScreen: React.FC<HomeScreenProps> = ({ navigation }) => {
   const [activeTab, setActiveTab] = useState("Tab1");
-  const handleCardPress = (item: {
-    title: string;
-    image: any;
-    prix: string;
-    editor: string;
-    developer: string;
-    date: string;
-    genre: string;
-    os: string;
-    processor: string;
-    memory: string;
-    graphics: string;
-    storage: string;
-    directX: string;
-    additionalNote: string;
-    screen: string;
-    promo: string;
-    year: string;
-  }) => {
-    console.log("Card pressed:", item);
-  };
+
 
   return (
     <ScrollView style={commonStyles.containerHomePage}>
